requirements.txt
#Django==3.1.3
Django-Crispy-Forms==1.9.0
Pillow==8.0.1
